fix #180 "preserved tokens" inside comments make the CSS compressor fail.
improvements:

-) additionally if on of the tokens to be preserved is not followed by a
terminator, the search loop just started over to search from the start
of the css text, finding the terminator of the first token.
This was fixed by breaking out the loop if no terminator can be found.

-) A token is skipped if its position is wihin an area already copied over
to the resulting string buffer.

-) However, the most important is, to search for comments prior to search
for any token to preserve. This seams reasonable since comments do not
contain any CSS information. They are just - comments.